Further improve docker setup
Use a python builder step to reduce the size of the final image.
Builder uses python:2-stretch to build dependenceis in a virtualenv.
Subsequent images are based on python:2-slim-stretch and copy the venv.
This technique is borrowed from Giorgos' work on Kitsune.

This rearranging also means that prod static assets are not built, and
prod data is not re-fetched (only the dev database) for local dev images.